u-boot-xlnx: Move all u-boot FIT image (blob) implementation to here
The implementation appeared to have been split between meta-som and
meta-xilinx-core.  The core implementation should be generic and in
meta-xilinx-core.  Specifics for the som board (or other boards) will
be captured in the board specific layers.

Note: This code may be moved in the future as part of a consolidation
of device tree related items.

Signed-off-by: Mark Hatle <mhatle@xilinx.com>