xilinx-qemu: Use addto_recipe_sysroot task
To populate the QEMU binaries into sysroot the populate_sysroot task
isn't enough for qemu-xilinx-helper-native, we need to use addto_recipe_sysroot task.
This patch will correct the mapping.

u-boot-zynq-uenv failing to add DEPENDS when EXTRA_IMAGEDEPENDS has
recipe:task. This patch will split the recipe name and add it to
DEPENDS.

Signed-off-by: Raju Kumar Pothuraju <raju.kumar-pothuraju@xilinx.com>

While unusual, the PR is updated, as the qemu and qemu helper need to have
the same version/pr or the path logic doesn't work properly.

Signed-off-by: Mark Hatle <mhatle@xilinx.com>