u-boot-xlnx-blob.inc: Move to using UBOOT_IMAGE_BLOB_DEFAULT
UBOOT_IMAGE_BLOB may need a machine (override) specific default value, but we
always want the user to be above to override this easily.  Adding the new
UBOOT_IMAGE_BLOB_DEFAULT allows us to set the default in a BSP layer, while
continuing to make the UBOOT_IMAGE_BLOB setting easy for the user to adjust.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>