libdfx_1.0.bb: Manually specify that 'libdfx' package is being replaced
The commit a64ff0bf015adcefbe691a94ac3e8ecf1a2ba911 updated the upstream
source to include correct soname/soversion.  Packaging was then modified
to follow this example.  The end result was that 'libdfx' was replaced by
the automatic "libdfx1.0" package.  This could result in:

Error: Transaction test error:
  file /usr/lib/libdfx.so from install of libdfx-dev-1.0-r0.1.cortexa72_cortexa53 conflicts with file from package libdfx-1.0-r0.0.cortexa72_cortexa53

The cause was that libdfx wasn't upgraded when libdfx-dev was being
installed.  To resolve this we use RREPLACES, may look strange but
this will instruct the system that "libdfx" is to be replaced by the
dynamically generated package based on the soname.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>