qemu-xilinx cleanup
Refactor qemuboot-xilinx.bbclass and machine-xilinx-qemu.inc.  Following
the example of the core version of qemu.bbclass and match inc file.

Refactor some of the functions to make them more readable.  Specifically
around qemu_rootfs_params.  Code may not be as compact, but it'll be
easier to read (and update) in the future.

Add mising qemu-system-native PREFERRED_PROVIDER (and recipe PROVIDE).
This resolves the issue where occasionally you use runqemu and the qemu
binary appears to be missing.

*-generic.conf: Remove overrides that are the same as the default qemu
settings.

Remove :append operations on QB_OPT_APPEND, this allows easy override
when someone includes a generic machine.

Signed-off-by: John Toomey <jtoomey@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>