dfx-mgr: new SRCREV, backport 22.2 patches
This is a cherry-pick of all patches in the master, i.e.
the xlnx_rel_v2022.1 branch has the same code as the master.
The commit ID’s are different, because of an unfortunate
merge commit in the dfx-mgr repo.

Signed-off-by: Sergei Korneichuk <sergei.korneichuk@amd.com>
(cherry picked from commit 13e1a00011e48dc20dae3e7f452590a3789a5d7f)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>