dt-processor.sh: Move BASE_TMPDIR from = to ?=
The build directory may be on an NFS server, while it's easy for a user to
modify the BASE_TMPDIR to a local device, this is difficult for an automated
build to do via another configuration file.

Switching to ?= will allow site.conf (or a similar file) to set a custom value
for BASE_TMPDIR to a local disk.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>