machine-xilinx-default.inc: Add WKS_FILE_DEPENDS based on IMAGE_BOOT_FILES
When the files pointed to by IMAGE_BOOT_FILES change, we need to let the
wic system know a new rootfs is required.  Do this via WKS_FILE_DEPENDS.

A number of automatic ones have been added to the system, if additional
files are added (or ones that use different names) it is up to the
machine.conf file [or configuration] to add the necessary dependencies.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>