image-wic-utils.bbclass: Support subdirectory parsing
Syntax for IMAGE_BOOT_FILES is <source>;<dest> where either or both
source can include a directory.  In the case of:

  devicetree/*.dtbo;devicetree/

The system will copy all of the contents (*.dtbo) from devicetree directory
and place it into a corresponding devicetree directory (using the same
filename) in the boot partition.

The code wasn't paying attention to this directory syntax, which caused
boot configuration errors.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>