machine-xilinx-default.inc: Enable both bootimg and rootfs image /boot
IMAGE_BOOT_FILES and WKS_FILE_DEPENDS are used in the bootimg technique.
THis ensures that an bootimg will continue to work in the future.

The MACHINE_ESSENTIAL_EXTRA_RDEPENDS is also filled out based on the
same IMAGE_BOOT_FILES.  This ensures that the packages used to generate
the filesystem's /boot partition is matching and full.

The end result of both approaches is slightly different, but the rootfs
method must be used for on-target package upgrade to work properly.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>