machine: Refactor mali400 support
Define a new MACHINE_FEATURE called mali400 that is used to determine if the
hardware has a mali400.

Define a new DISTRO_FEATURE called libmali, as well as one called lima to
indicated which configuration to build.

Both the MACHINE_FEATURE and DISTRO_FEATURE must be enabled to activate
the new code.  This means any recipe that gets libmali-xlnx is expected
to be:
   PACKAGE_ARCH = "${MACHINE_ARCH}"

Signed-off-by: Mark Hatle <mark.hatle@amd.com>