standalone gcc-runtime_%.bbappend: Add workaround for getentropy and related
getentropy and the related arc4random function are defined in unistd.h, they
appear to be defined in some iterations of libgloss.  However the defintion
is not enough in the general case, which causes C++ linkage to fail with:

  libstdc++.a(random.o): in function `std::(anonymous namespace)::__libc_getentropy(void*)':
  (.text._ZNSt12_GLOBAL__N_117__libc_getentropyEPv+0x14): *undefined reference to `getentropy'*

This change doesn't remove getentropy or arc4random, but does instruct
libstdc++ to not use either function.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>