meta-xilinx-setup: Rework dt-processor.sh to generate a machine and configs
lop-microblaze-yocto.dts has been modified to generate the file in a slightly
different format.  There is no need for an intermediate representation for
the various microblaze tunings, instead use fixed names for pmc, psm and pmu
configurations.

dt-processor.sh was reworked to generate a machine.conf file and shrink the
amount of information that ends up in the local.conf file.

As part of this work the -m option has been renamed to -t.  This option is
rarely used, so should not impact many users.

A new -m option is implemented to allow the user to manually name the
generated machine.conf file.  Otherwise, it will discover the name from the
system device tree using the root 'compatible' object.

The tool will also print more useful diagnostics to the screen, this should
help ensure that the correct machine is being targeted by the actions and
improve early system debuggability.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>