xilinx-standalone.inc: Sync with poky distro settings
Certain settings appear to have an overall affect on recipe hashes. Use
petalinux (poky) distro as an example for base settings to ensure native
items built in one config SHOULD also be re-usable in the bareemetal config.

This has a side effect of NOT disabling static binaries in baremetal, but
this seems like a reasonable default, since there really isn't a concept
of dynamically loaded libraries.  If necessary, the user can still override
this, but will need to specifically select class-target to do so.

Similarly SECURITY_CFLAGS and SECURITY_LDFLAGS are available by empty by
default.  The default cflags/ldflags for baremetal usage are likely
different from a Linux system, so can be used in the future if requried.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>