qemu-xilinx-system-native: Move pmu-rom enablement to recipe
Setting the pmu-rom being required by QEMU within the machine.conf will cause
the recipe's hash to change, which can cause a ripple effect causing other
things to rebuild.  Moving this to the recipe, and switching it on based
on what licenses the user allows will avoid this conflict.

Additionally add a warning message to the xilinx-vars.bbclass if the user has
not enabled the license.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>