machine-xilinx-default: Implement default FIT_CONF_DEFAULT_DTB
When using a system device tree build there will be more then one device-tree
that is constructed.  If a fit image is constructed, then an DTB needs to be
selected.  We use the CONFIG_DTFILE as the default, which should inherit
from the SYSTEM_DTFILE.  The default for SYSTEM_DTFILE is empty as a fallback
which should restore the default behavior of first DTB found is the one used.

This also has ramifications for other recipes that use DTB files and unifies
everything.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>