machine-xilinx-default.inc: defining default dtb for fitimage
When we tried to create fitimage it will pick the dtb from linux recipe-sysroot path.
There are multiple dtb files in linux sysroot in sdt case and we did not specified any default dtb file to package in fitimage and it was picking baremetal dtb file which is leading to kernel panic.

There is a variable in yocto to specify the default dtb file to pick for fitimage FIT_CONF_DEFAULT_DTB.
If we specify the above then the fitimage is generated properly and able to boot successfully.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>