machine-xilinx-default.inc: Set FIT_CONF_DEFAULT_DTB only for SDT flow
system.dtb is symlink to actual dtb file which is not present in linux
sysroot path. If specified dtb name for FIT_CONF_DEFAULT_DTB not present
fitimage.bbclass will not set default configuration variable.

To fix this, setting FIT_CONF_DEFAULT_DTB to empty if CONFIG_DTFILE not
set(xsct flow) so that which dtb comes first that will be the default
configuration in fitimage creation.

Signed-off-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>