zynqmp-generic: Emulate FSBL hand-off for qemu boot
Qemu boots by loading ATF and jumping to it directory.  A recent change
adds a verification that the data structure passed by the FSBL is correct
in a 'mode 5' (SD card) boot.  If it's not valid ATF will fail to continue
to boot.  (JTAG, 'mode 0' boot will proceed but other system behavior
changes.)

In order to emulate the FSBL handoff to ATF, we need to populate a data
structure in fffc0000, and tell ATF to look at that address.  See the
comments for more information.  This also ends up hard-coding the u-boot
address.  If it changes booting will fail.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>