device-tree: Move from file CONFIG_DTFILE access to SRC_URI
Move from declaring DT_FILES_PATH to the CONFIG_DTFILE dirname, to the
default where it is set to S.  This adds the requirement that CONFIG_DTFILE
_must_ be self contained (or preprocessed).

This prevents the system from compiling any other dts (or dtsi) files
located in that directory, to only the ones we've specifically requested.

Additionally this simplifies the check_devicetree_variables, as hashing
and file validation are now done for us by the fetchers.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>