xilinx-bootbin: Refactor create_bif functions
Remove the biffunc dict which was used to map arch to a specific bif
creation function and replace with simple if/elif/else logic. The
default bif file functions are defined for zynq/zynqmp and versal as
well as and empty default function which could be redefined in a
bbappend with new logic or mapped back to one of the existing
functions.

Signed-off-by: John Toomey <john.toomey@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>