weston: Add Weston 9.0.0 for libmali, disable Weston 10.0.2 with libmali
libmali only provides GL ES 2 interfaces, but Weston 10.0.2 requires GL ES 3
interfaces.  Disble 10.0.2 version in this case, and fall back to 9.0.0.

The 9.0.0 version is based on the last 9.0.0 supported Yocto Project version.
One minor change was required, in order to support on-target upgrade from the
unmodified 10.0.2 to the libmali 9.0.0 requires a way to enabel PE set to 1.

However if the 9.0.0 version doesn't use libmali, we don't want the PE set
so that 10.0.2 remains the superior version.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>