meta-xilinx-core: qemu: Add additional RPROVIDES for alternative QEMUs
The alternative QEMU needs a way to provide compatible package names for
dynamic packages, as well as the static ones it was already supporting.

Do this by copying the package split routine and adjusting it to
add these additional rprovides as necessary.

Additionally, we need to tell the system any dynamic packages will be
available so that dependencies can be correctly specified.  The code
change above happens well after parse, so there is no direct way
for it to work.  We fix this by duplicating the wild carding that is
already in the PACKAGES_DYNAMIC.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>