meta-xilinx-core: conf/layer.conf: Skip xrt in dependency hashing
xrt API is stable, but we provide different versions if there is an ai-engine
or not.  Skip the hash calculations to allow dependent packages to work with
both versions.

This resolves a problem where xrt w/ and w/o ai-engine support causes vart
and thus vitis-ai to rebuild multiple times.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>