esw-conf: Move to embeddedsw and use BBCLASSEXTEND
Make sure that we have native and nativesdk versions of esw-conf.

Add a dependency on lopper and xilinx-lops, as presumably users of this
recipe will need to use lopper to process the configurations.  This will
allow someone to use oe-run-native with esw-conf-native to use lopper
to transform system to domain specific device trees and their associated
configurations.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>