various machines: Change qemu serial port configuration for nanbield
As of Yocto Project nanbield, if a -serial is pass in QB_OPT_APPEND the
runqemu may add additional null entries (padding) or simply skip further
setup.

Due to this, emulated platforms where the first hardware serial port
is not used in Linux need specific "-serial null" settings, but with
the new rules also need the real serial port defined (this can no
longer be assumed.)

The downside of this new approach is all consoles must be defined, so
we can not rely on runqemu adding "-serial mon:<device>", as now
assume the primary usage is stdio (nographic/serial port) settings.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>