dfx_user_dts.bbclass: Fix build issue with bin and dtsi as input
In dfx_user_dts.bbclass fix build issue with bin and dtsi as input file
to firmware recipe. In case of bin file as input bootgen operation is
not performed and directly copy from ${S} directory. Also remove check
condition if dtbo with firmware-name doesn't exist which is not
required in case of bin file as input.

Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>