meta-xilinx-core: virtualization_layer: qemu
Rework the meta-virtualization qemu extensions implemented in the
meta-xilinx-core layer.  It was found that is a bug in the langdale
meta-virtualization.  Due to this, we need to use BBMASK to exclude
the changes from that layer, and then repeat them, but only for
qemu 7.1.  qemu 8.1 work is taken from the current master version
of meta-virtualization.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>