standalone: Add support for Versal Net specific CMAKE variables
CMAKE_MACHINE and CMAKE_SYSTEM_PROCESSOR values are different for Versal Net
than that of Versal. Update the same using the SOC_VARIANT as the SOC_FAMILY is
same for both.

In addition, update the SOC specific flags (defined in-house to segragate execution
of some part of the boot code and the drivers) for versal and versal-net.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>