device-tree: Replace symlink with cp
It is not currently possible to create symlinks in the /boot partition
due to the use of the vfat filesystem and this causes an issue when
trying to update the dtb using rpm packages which contain a symlink.

Work around the issue by copying the dtb file to the symlink location.

Signed-off-by: John Toomey <john.toomey@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>
(cherry picked from commit 6ddde9a20775549f1434cba6006f84db1a223cb4)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>