u-boot-xlnx-2023.2.inc: space is missing for SRC_URI:append
when user changes any config in u-boot menu config then
we are saving that in user-<timestamp>.cfg.
While adding that user*.cfg to SRC_URI space is missing
and it was treated as different file and getting compilation issue.
The space missed in below line
meta-xilinx/meta-xilinx-core/recipes-bsp/u-boot/u-boot-xlnx-2023.2.inc#L19.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>