esw.bbclass: Cleanup
PV defaults to part of the FILE name

ESW_VER defaults to the same part of the FILE name

Both are intended to be overriden independently by the recipe, if necessary.
So remove the default value as it will already work as intended and avoid
future confusion.

Also cleanup the comment to explain why the remove is necessary.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>