libomxil-xlnx: Remove obsolete versions
There is no reason to keep the older versions, as the current version remains
the same API, just with fixes.

Note before the version was set to 1.0.0, but the various version files inside
the source indicate it's really 1.1.2.  Adjust the naming to match and make
it clear this is the xilinx fork and which release.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>