qemu: Update from older versions to 2024.1 - 8.1.0 version
The 8.1.x integration is a forward port from poky as of commit
4bb222e0d71a4cb159b8a4f1a90b65b1af32ac10.  We need the 8.1.x version to
match the AMD qemu-xlnx fork version.

Remove all of the older 2022.1 - 2023.2 versions.  No reason to keep these
old versions, as the latest will do everything the older one can.

Note, some of the devicetrees may have changed names, requiring BSPs with
integrated PS and PMU/PMC DTBs to be changed.

This version corrects the QMP integration with the Yocto Project ptest
runner.  There may still be issues with the ptest runner, but qemu itself
should now be compatible.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>