zynqmp-generic.conf: Fix qemu pmu firmware path
The path specified is the internal firmware path.  This could be in a the
current deploy directory, a multiconfig deploy directory or some other place
directly specified by the user.

When using the sstate-cache, the build of the component in the a multiconfig
directory can be skipped, causing an error to find the component.

Instead we want to use OUR deploy directory, which will always have a copy
of the pmu firmware, no matter how it was constructed or packaged.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>