python3-bokeh: add and update
Add (move) the python3-bokeh recipe to here from meta-jupyter while updating
from 2.4.2 → 3.5.1. Bokeh is not a requirement or dependency for jupyter,
therefore it does not need to be stored in that layer.

Bokeh is used by:
	- meta-kria/recipes-utils/kria-dashboard/kria-dashboard.bb
	- meta-system-controller/recipes-utils/pmtool/pmtool.bb

Both of those layers have dependencies on meta-xilinx-core, so moving the
recipe here will still make it available to both.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>