meta-xilinx-standalone-sdt: sdt-artifacts: Rework to implment SDT_URI
Rename to a version-less component.  We're not tied to a specific version
of Vivado for this.

Replace SDT_PROT, SDT_PATH, SDNAME and SDT_FILENAME with:

SDT_URI - URI for the system device tree download
SDT_URI[sha256sum] - sha256 for the file refernced in SDT_URI
SDT_URI[S] - Specific extracted S(OURCE) dir (optional)

If SDT_URI is a local directory, SDT_URI[sha256sum] is not required.

SDT_URI[S] must be set if the component is extracted into a subdirectory of
${WORKDIR}.  Most configurations will likely need this.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>