meta-xilinx-demos: meta-ros2-jazzy: fix orocos-kdl-vendor build
The orocos-kdl-vendor build was failing due to being unable to find the
requested commit hash from the specified branch (release/jazzy/
orocos_kdl_vendor). The originally-generated recipe was mostly correct, but
was overridden by a bbappend in upstream meta-ros/meta-ros2-jazzy which
replaced the hash with one from the release/rolling/python_orocos_kdl_vendor/
0.4.1-1 branch.

The jazzy branch of meta-ros2 was created from the rolling release branch,
which is why the orocos-kdl recipe had an override which specified a commit
from the rolling branch of orocos-kdl. This wasn't updated to point to a
commit on the jazzy branch of orocos-kdl when the jazzy branch was created.

Update the SRCREV to point to a release on the release/jazzy/orocos_kdl_vendor
branch of orocos-kdl. Also, update to the latest commit on this branch.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>