meta-xilinx-demos: meta-ros2-jazzy: mcap-vendor: fix build
The rosbag2/mcap-vendor build was failing due to being unable to
find the requested commit hash.

The originally-generated recipe was mostly correct, but was overridden by a
bbappend, in upstream meta-ros/meta-ros2-jazzy, which replaced the hash with
one from the release/rolling/mcap_vendor/0.24.0-1 branch.

The jazzy branch of meta-ros2 was created from the rolling release branch,
which is why the rosbag2/mcap-vendor recipe had an override which specified
a commit from the rolling branch of rosbag2-release. This wasn't updated to
point to a commit on the jazzy branch when the jazzy branch was created.

Update the SRCREV to point to a release on the release/jazzy/mcap_vendor
branch of rosbag2-release Also, update to the latest commit on this branch.

Additionally the build was failing because it was not able to find necessary
header files that were provided in the source tree.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>