meta-xilinx-core: xrt: Rework XRT AIE dependencies
The older recipes, merge xrt_% into the specific versioned items to ensure
that they continue to work the way they have in the past.

For the current xrt version, we remove the MACHINE dependencies and allow
AIE to be selected via a PACKAGECONFIG instead.  This will permit a single
xrt to work across a larger set of aarch64 machines.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>