xilinx-standalone: Add workaround for warning about embedded buildpaths
Normally the presence of buildpaths indicates a problem with a library or
executable.  This is resolved by the system passing in prefix mapping
values to ensure all paths are target oriented.  For some reason,
only in baremetal configurations, part of the debug paths are NOT
being remapped.  These paths are not actually used anywhere:

 The Directory Table (offset 0x22, lines 6, columns 1):
  Entry Name
  0     (indirect line string, offset: 0x74): /scratch/mhatle/git/2024.2/build/tmp-zynqmp-zcu104-sdt-full-cortexa53-fsbl/work/cortexa53-xilinx-elf/csudma/2024.2+git/build
  1     (indirect line string, offset: 0x102): /usr/include/machine
  2     (indirect line string, offset: 0x117): /usr/include/sys
  3     (indirect line string, offset: 0x128): /usr/lib/aarch64-xilinx-elf/gcc/aarch64-xilinx-elf/13.3.0/include
  4     (indirect line string, offset: 0x16a): /usr/include
  5     (indirect line string, offset: 0x177): /usr/src/debug/csudma/2024.2+git/XilinxProcessorIPLib/drivers/csudma/src

 The File Name Table (offset 0x40, lines 12, columns 2):
  Entry Dir     Name
  0     5       (indirect line string, offset: 0x1c0): xcsudma.c
  1     1       (indirect line string, offset: 0x1ca): _default_types.h
  2     2       (indirect line string, offset: 0x1db): _stdint.h
  3     3       (indirect line string, offset: 0x1e5): stddef.h
  4     4       (indirect line string, offset: 0x1ee): xil_types.h
  5     5       (indirect line string, offset: 0x1fa): xcsudma.h
  6     4       (indirect line string, offset: 0x204): xil_util.h
  7     4       (indirect line string, offset: 0x20f): string.h
  8     4       (indirect line string, offset: 0x218): xil_cache.h
  9     4       (indirect line string, offset: 0x224): xil_assert.h
  10    5       (indirect line string, offset: 0x231): xcsudma.c
  11    4       (indirect line string, offset: 0x23b): xil_io.h

In the above example, entry '0' in the directory table is tiggering
the the buildpath QA warning.  However in the file name table the
path is never referred to.  Also I have verified that all of the path
remapping arguments are being passed in correctly during compilation
and linking.

Based on this, it appears to be a valid workaround to simply disable
the test while we determine why the invalid enty is present.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>