distro/xilinx-standalone: Add inclusion to petalinux-mirrors.conf
By default PetaLinux uses the own-mirror class to access cached version of
the software.  While this distribution configuration is not PetaLinux, nor
should it require PetaLinux to operating properly, including the same mirror
configuration will unify the Linux and non-Linux (baremetal) configurations.

This resolves an issue with BB_NO_NETWORK set to 1, and a local disk mirror
is enabled.

Uses 'include' instead of 'require' so no error is generated if meta-petalinux
(which it not required) is not present in the user's configuration.  In this
case behavior reverts to default and any mirror configurations also return
to default behavior.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>