meta-xilinx-standalone: Rework embeddedsw integration
Introduce a new embeddedsw-source recipe that is a single point for unpack,
and patch of the sources.  This will allow someone to universally patch
(via a bbappend or similar) the esw sources.

As part of this work, we transfered the various patches for fsbl, pmu, plm,
and psm to the universe source recipe.

Most of this is transparent to any recipe using xlnx-embeddedsw.  The class
automaticaly detects if the recipe is a user our the special embeddedsw-source
recipe and enables the correct behavior.

Since the esw sources expect the S and B to be the same, the cmake objects
get written into S, we are not able to follow the gcc-source example.
Instead recipes that use the xlnx-embeddedsw class will hard link or copy
the source files to their working directory.  This avoid the penalty for
fetch, unpack and patch.  Note these recipes may also add their own
SRC_URI which WILL be processed as normal.  This is often used to
introduce special tcl scripts.  See the do_copy_shared_src task injected
by the xlnx-embeddedsw.bbclass for technical details.

Additionally, correct some of the existing patch Upstream-Status entries
to conform to current YP guidelines.

For 2024.2 introduce a special Xil_Assert fix for the current toolchain
which is more strict on type conversion..  See the patch itself for
more details.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>